+++ frontend/src/App.tsx
@@ -3,10 +3,10 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import MobileLayout from './common/components/MobileLayout/MobileLayout';
 import { PAGE_URL } from './common/constants/url';
 import Booking from './pages/booking/Booking';
+import CreatedMentoring from './pages/CreatedMentoring/CreatedMentoring';
 import Detail from './pages/detail/Detail';
 import Home from './pages/home/Home';
 import MentoringCreate from './pages/mentoringCreate/MentoringCreate';
-import CreatedMentoring from './pages/myPage/components/CreatedMentoring/CreatedMentoring';
 import MyPage from './pages/myPage/MyPage';
 
 const router = createBrowserRouter([
